+++ lib/end_of_life/cli.rb
@@ -2,7 +2,7 @@ module EndOfLife
   class CLI
     def call(argv)
       parse_options(argv)
-        .then { |options| execute_command(options) }
+        .then { |options| execute_command(argv, options) }
     end
 
     private
@@ -11,7 +11,7 @@ module EndOfLife
       Options.from(argv)
     end
 
-    def execute_command(options)
+    def execute_command(argv, options)
       case options[:command]
       in :help
         puts options[:parser]
@@ -21,6 +21,8 @@ module EndOfLife
         abort options[:error]
       in :scan
         Scanner.scan(options)
+      in :check
+        Check.run(argv, options)
       end
     end
   end