+++ lib/end_of_life/helpers/text.rb
@@ -1,0 +1,11 @@
+module EndOfLife
+  module Helpers::Text
+    def pluralize(count, singular, plural = nil)
+      if count == 1
+        "#{count} #{singular}"
+      else
+        "#{count} #{plural || "#{singular}s"}"
+      end
+    end
+  end
+end