+++ lib/end_of_life/product/release.rb
@@ -3,6 +3,17 @@ module EndOfLife
     class Release < Data.define(:product, :version, :eol_date)
       include Comparable
 
+      def self.parse!(string)
+        product, version = string.split("@", 2)
+        raise ArgumentError, "Invalid product release format: #{string}" if product.to_s.empty? || version.to_s.empty?
+
+        begin
+          new(product:, version:)
+        rescue ArgumentError
+          raise ArgumentError, "Malformed version number string: #{version}"
+        end
+      end
+
       def self.ruby(version, eol_date: nil) = new(product: "ruby", version:, eol_date:)
 
       def initialize(product:, version:, eol_date: nil)
@@ -17,6 +28,16 @@ module EndOfLife
         else
           self <= product.latest_eol_release(at: at)
         end
+      end
+
+      def supported?(...) = !eol?(...)
+
+      def latest_cycle_release
+        product.all_releases.filter { |r| r.version.to_s.start_with?(cycle_version.to_s) }.max
+      end
+
+      def cycle_version
+        Gem::Version.new(version.segments.first(2).join("."))
       end
 
       ZERO = Gem::Version.new("0")